+++ authpass/lib/l10n/app_localizations_de.dart
@@ -429,6 +429,11 @@ class AppLocalizationsDe extends AppLocalizations {
   }
 
   @override
+  String passwordScore(Object score) {
+    return 'Strength: ${score} of 4';
+  }
+
+  @override
   String unexpectedError(String error) {
     return 'Unerwarteter Fehler: ${error}';
   }