+++ authpass/lib/ui/screens/cloud/cloud_mailbox.dart
@@ -12,10 +12,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-import 'package:kdbx/kdbx.dart';
-import 'package:provider/provider.dart';
-
 import 'package:logging/logging.dart';
+import 'package:provider/provider.dart';
 
 final _logger = Logger('cloud_mailbox');