+++ authpass/CHANGELOG.md
@@ -3,6 +3,7 @@
 * Android: improve autofill input field detection for usernames.
 * Improve locking with biometric storage.
   (Files can now be closed/locked without removing saved master passwords)
+* File Create: Show password strength #143
 * Password generator: include accented characters in 'Umlauts'
 * French translations.